+++ editors/zed/languages/ppphp/indents.scm
@@ -1,0 +1,10 @@
+[
+  (compound_statement)
+  (declaration_list)
+  (enum_declaration_list)
+  (arguments)
+  (formal_parameters)
+  (array_creation_expression)
+  (switch_block)
+  (match_block)
+] @indent